added spider that scrapes göteborgs nation for the phone number to book lunch

+++ spider/nationslunchspider/pipelines.py
@@ -11,10 +11,15 @@ import json
 class CustomPipeline(object):
 
     def open_spider(self, spider):
-        self.file = open('c:/Users/Markus/desktop/projects/nationslunch/spider/nationslunchspider/output/images.jl', 'w')
-        self.file.truncate()
+        if spider.name == 'Nationsspider':
+            self.file = open('c:/Users/Markus/desktop/projects/nationslunch/spider/nationslunchspider/output/images.jl', 'w')
+            self.file.truncate()
     
+        if spider.name == 'gspider':
+            self.file = open('../output/phone.jl','w')
+            self.file.truncate()
 
+       
     def close_spider(self,spider):
         self.file.close()
